feat: implement persistent layout
