+++ src/components/BlogPost.tsx
@@ -1,24 +1,12 @@
 import React, { useState, useEffect, Suspense, lazy } from 'react'
 import { helmetJsonLdProp } from 'react-schemaorg'
 import { BlogPosting } from 'schema-dts'
-import {
-  Container,
-  CssBaseline,
-  Box,
-  Grid,
-  Typography,
-  Link,
-  Breadcrumbs,
-  Fab,
-  IconButton,
-  Tooltip,
-} from '@mui/material'
+import { Box, Typography, Link, Breadcrumbs, Fab, IconButton, Tooltip } from '@mui/material'
 import KeyboardArrowUpIcon from '@mui/icons-material/KeyboardArrowUp'
 import BookmarkIcon from '@mui/icons-material/Bookmark'
 import BookmarkBorderIcon from '@mui/icons-material/BookmarkBorder'
 import { Helmet } from 'react-helmet'
 import { Link as RouterLink, useLocation } from 'react-router-dom'
-import Header from './Header'
 import LoadingSkeleton from './LoadingSkeleton'
 import withCanonical from './WithCanonical'
 import { PROFILE } from '../constants/profile'
@@ -28,10 +16,6 @@ import profileImage from '../assets/raymond-perez.jpg'
 import SocialShareButtons from './SocialShareButtons'
 import { posts } from '../constants/posts'
 
-// Lazy load below-the-fold components
-const ProfileCard = lazy(() => import('./ProfileCard'))
-const RecentPosts = lazy(() => import('./RecentPosts'))
-const BookmarkedPosts = lazy(() => import('./BookmarkedPosts'))
 const AuthorBio = lazy(() => import('./AuthorBio'))
 const TableOfContents = lazy(() => import('./TableOfContents'))
 
@@ -116,140 +100,112 @@ const BlogPost: React.FC<BlogPostProps> = ({ title, author, date, children }) =>
           {title} - {PROFILE.name} - {PROFILE.role}
         </title>
       </Helmet>
-      <Container maxWidth={false}>
-        <CssBaseline />
-        <Box my={2}>
-          <Grid container spacing={2} direction='row-reverse' alignItems='flex-start'>
-            <Grid item xs={12} lg={4}>
-              <Header />
-              <Suspense fallback={<LoadingSkeleton testId='profile-card' />}>
-                <ProfileCard image={profileImage} name={PROFILE.name} role={PROFILE.role} />
-              </Suspense>
-              <Suspense fallback={<LoadingSkeleton testId='bookmarked-posts' />}>
-                <BookmarkedPosts />
-              </Suspense>
-              <Suspense fallback={<LoadingSkeleton testId='recent-posts' />}>
-                <RecentPosts />
-              </Suspense>
-            </Grid>
-            <Grid item xs={12} lg={8}>
-              <Box mb={2}>
-                <Breadcrumbs aria-label='breadcrumb'>
-                  <Link color='inherit' href='/'>
-                    Home
-                  </Link>
-                  <Typography color='textPrimary'>{title}</Typography>
-                </Breadcrumbs>
-              </Box>
-              <Box display='flex' alignItems='center' justifyContent='space-between' mb={2}>
-                <Typography variant='h2' component='h1' sx={{ flexGrow: 1 }}>
-                  {title}
-                </Typography>
-                <Tooltip title={isCurrentlyBookmarked ? 'Remove Bookmark' : 'Bookmark Article'}>
-                  <IconButton
-                    onClick={handleBookmarkClick}
-                    aria-label={isCurrentlyBookmarked ? 'remove bookmark' : 'bookmark article'}
-                    sx={{
-                      color: isCurrentlyBookmarked ? 'primary.main' : 'text.secondary',
-                      ml: 2,
-                      '&:hover': {
-                        color: 'primary.main',
-                        bgcolor: 'action.hover',
-                      },
-                    }}
-                  >
-                    {isCurrentlyBookmarked ? <BookmarkIcon /> : <BookmarkBorderIcon />}
-                  </IconButton>
-                </Tooltip>
-              </Box>
-              <SocialShareButtons title={title} />
-              <Box mb={2}>
-                <Typography variant='body2' color='text.secondary'>
-                  {readingTimeDisplay}
-                </Typography>
-              </Box>
-              <Suspense fallback={<LoadingSkeleton testId='table-of-contents' />}>
-                <TableOfContents />
-              </Suspense>
-              {children}
-              <Suspense fallback={<LoadingSkeleton testId='author-bio' />}>
-                <AuthorBio />
-              </Suspense>
-              {/* Previous/Next Navigation */}
-              {(prevPost || nextPost) && (
-                <Box
-                  mt={6}
-                  mb={2}
-                  display='flex'
-                  justifyContent='space-between'
-                  alignItems='center'
-                >
-                  {prevPost ? (
-                    <Box>
-                      <Typography variant='caption' color='text.secondary'>
-                        Previous
-                      </Typography>
-                      <Link
-                        component={RouterLink}
-                        to={prevPost.path}
-                        underline='hover'
-                        color='primary'
-                        sx={{ display: 'block', fontWeight: 500 }}
-                      >
-                        {prevPost.title}
-                      </Link>
-                    </Box>
-                  ) : (
-                    <span />
-                  )}
-                  {nextPost ? (
-                    <Box textAlign='right'>
-                      <Typography variant='caption' color='text.secondary'>
-                        Next
-                      </Typography>
-                      <Link
-                        component={RouterLink}
-                        to={nextPost.path}
-                        underline='hover'
-                        color='primary'
-                        sx={{ display: 'block', fontWeight: 500 }}
-                      >
-                        {nextPost.title}
-                      </Link>
-                    </Box>
-                  ) : (
-                    <span />
-                  )}
-                </Box>
-              )}
-            </Grid>
-          </Grid>
-        </Box>
-        {showBackToTop && (
-          <Fab
-            size='small'
-            onClick={scrollToTop}
-            aria-label='back to top'
+      <Box mb={2}>
+        <Breadcrumbs aria-label='breadcrumb'>
+          <Link component={RouterLink} color='inherit' to='/'>
+            Home
+          </Link>
+          <Typography color='textPrimary'>{title}</Typography>
+        </Breadcrumbs>
+      </Box>
+      <Box display='flex' alignItems='center' justifyContent='space-between' mb={2}>
+        <Typography variant='h2' component='h1' sx={{ flexGrow: 1 }}>
+          {title}
+        </Typography>
+        <Tooltip title={isCurrentlyBookmarked ? 'Remove Bookmark' : 'Bookmark Article'}>
+          <IconButton
+            onClick={handleBookmarkClick}
+            aria-label={isCurrentlyBookmarked ? 'remove bookmark' : 'bookmark article'}
             sx={{
-              position: 'fixed',
-              bottom: 16,
-              right: 16,
-              bgcolor: 'background.paper',
-              color: 'text.secondary',
-              border: '1px solid',
-              borderColor: 'divider',
-              opacity: 0.8,
+              color: isCurrentlyBookmarked ? 'primary.main' : 'text.secondary',
+              ml: 2,
               '&:hover': {
-                opacity: 1,
+                color: 'primary.main',
                 bgcolor: 'action.hover',
-                borderColor: 'primary.main',
               },
             }}
           >
-            <KeyboardArrowUpIcon />
-          </Fab>
-        )}
-      </Container>
+            {isCurrentlyBookmarked ? <BookmarkIcon /> : <BookmarkBorderIcon />}
+          </IconButton>
+        </Tooltip>
+      </Box>
+      <SocialShareButtons title={title} />
+      <Box mb={2}>
+        <Typography variant='body2' color='text.secondary'>
+          {readingTimeDisplay}
+        </Typography>
+      </Box>
+      <Suspense fallback={<LoadingSkeleton testId='table-of-contents' />}>
+        <TableOfContents />
+      </Suspense>
+      {children}
+      <Suspense fallback={<LoadingSkeleton testId='author-bio' />}>
+        <AuthorBio />
+      </Suspense>
+      {(prevPost || nextPost) && (
+        <Box mt={6} mb={2} display='flex' justifyContent='space-between' alignItems='center'>
+          {prevPost ? (
+            <Box>
+              <Typography variant='caption' color='text.secondary'>
+                Previous
+              </Typography>
+              <Link
+                component={RouterLink}
+                to={prevPost.path}
+                underline='hover'
+                color='primary'
+                sx={{ display: 'block', fontWeight: 500 }}
+              >
+                {prevPost.title}
+              </Link>
+            </Box>
+          ) : (
+            <span />
+          )}
+          {nextPost ? (
+            <Box textAlign='right'>
+              <Typography variant='caption' color='text.secondary'>
+                Next
+              </Typography>
+              <Link
+                component={RouterLink}
+                to={nextPost.path}
+                underline='hover'
+                color='primary'
+                sx={{ display: 'block', fontWeight: 500 }}
+              >
+                {nextPost.title}
+              </Link>
+            </Box>
+          ) : (
+            <span />
+          )}
+        </Box>
+      )}
+      {showBackToTop && (
+        <Fab
+          size='small'
+          onClick={scrollToTop}
+          aria-label='back to top'
+          sx={{
+            position: 'fixed',
+            bottom: 16,
+            right: 16,
+            bgcolor: 'background.paper',
+            color: 'text.secondary',
+            border: '1px solid',
+            borderColor: 'divider',
+            opacity: 0.8,
+            '&:hover': {
+              opacity: 1,
+              bgcolor: 'action.hover',
+              borderColor: 'primary.main',
+            },
+          }}
+        >
+          <KeyboardArrowUpIcon />
+        </Fab>
+      )}
     </React.Fragment>
   )
 }
